Use dynamic imports for wagmi providers to prevent SSR initialization errors
