+++ apps/web/src/providers-client.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import { WagmiProvider } from 'wagmi';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { getConfig } from '@/config/wagmi';
+
+const queryClient = new QueryClient();
+const config = getConfig();
+
+export function ProvidersClient({ children }: { children: React.ReactNode }) {
+  return (
+    <WagmiProvider config={config}>
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    </WagmiProvider>
+  );
+}
