+++ playwright-report/index.html
@@ -87,4 +87,4 @@ Error generating stack: `+l.message+`
     <div id='root'></div>
   </body>
 </html>
-<template id="playwrightReportBase64">data:application/zip;base64,UEsDBBQAAAgIAIZhk1xZNCEEwgAAAA0BAAALAAAAcmVwb3J0Lmpzb25Vj7luwzAQRH+F2JoQSMKQDPZp0yRd4GIjrRxG4oHlKk4g6N8N+SjSzbxi8GaFSIIDCoJfNw1VkOU9RAJvu65tjWmtaw8HDcPCKCEn8NYZ2x2bzjkNY5ipgv84aSicv6mXV4xPUgWlgl9BsuAM3mig30K90HArS/pXxxmnv1uqUyjlQfMEXnihTQMxZ9634WVPXu2ig8qLqAsGCemsrDPGmFjVyDmqPqcxnJsLfb4R/xA3cNKQy36j3v9G7L9CugtvV1BLAQI/AxQAAAgIAIZhk1xZNCEEwgAAAA0BAAALAAAAAAAAAAAAAAC0gQAAAAByZXBvcnQuanNvblBLBQYAAAAAAQABADkAAADrAAAAAAA=</template>
+<template id="playwrightReportBase64">data:application/zip;base64,UEsDBBQAAAgIAOBhk1ycr1h3xQAAABUBAAALAAAAcmVwb3J0Lmpzb25Vjztvg0AQhP/KaesTOkACfH3aNEkXuVjD4lzgHtpb4kSI/27hR+GpZr5iNLOCJ8EBBcGum4YsyPLpPIEt27ZpjGm7sulqDcPCKC4GsGVlqvpQdF19eEjD6GbKYL+OGhLHH+rlHf2TZEHJYFeQKDiDNRroL1EvNNzCEl7iOOP0f3N5cik9aJzACi+0aSDmyHs3vO3Oqn3woOIi6oJOXDirsjLGGJ/VyNGrPobRnYsLnT6If4kLOGqIab+T77899t8u3AdvV1BLAQI/AxQAAAgIAOBhk1ycr1h3xQAAABUBAAALAAAAAAAAAAAAAAC0gQAAAAByZXBvcnQuanNvblBLBQYAAAAAAQABADkAAADuAAAAAAA=</template>
