+++ test-results/results.json
@@ -134,8 +134,8 @@
     }
   ],
   "stats": {
-    "startTime": "2026-04-19T12:10:12.644Z",
-    "duration": 120178.722,
+    "startTime": "2026-04-19T12:13:01.683Z",
+    "duration": 120239.88399999999,
     "expected": 0,
     "skipped": 0,
     "unexpected": 0,
